+++ CHANGELOG.md
@@ -11,4 +11,5 @@
 ### Changed
 
 - Fix bug of deleting category. Now deletes all the category's balances instead of just the current.
+- Fix bug about the float value and comma char at the input number.